IdHelper (InfixExpr)

+++ adaptor/src/org/antlr/jetbrains/adaptor/parser/parsing/IdHelper.java
@@ -2,6 +2,8 @@ package org.antlr.jetbrains.adaptor.parser.parsing;
 import com.intellij.lang.PsiBuilder;
 import com.intellij.psi.tree.IElementType;
 import org.antlr.v4.runtime.ParserRuleContext;
+import org.antlr.v4.runtime.tree.ParseTree;
+import org.antlr.v4.runtime.tree.RuleNode;
 import org.antlr.v4.runtime.tree.TerminalNode;
 import org.jetbrains.plugins.scala.lang.ScalaLangParser;
 import org.jetbrains.plugins.scala.lang.parser.ScalaElementTypes;
@@ -15,6 +17,12 @@ public class IdHelper implements Helper {
     public void visitTerminal(TerminalNode node, PsiBuilder builder) {}
     @Override
     public void exitEveryRule(ParserRuleContext ctx, PsiBuilder.Marker marker, final Deque<PsiBuilder.Marker> markers) {
-        marker.drop();
+        RuleNode p = (RuleNode)ctx.getParent();
+        if (p.getRuleContext().getRuleIndex() == ScalaLangParser.RULE_infixExpr) {
+            marker.done(ScalaElementTypes.REFERENCE_EXPRESSION());
+        }
+        else {
+            marker.drop();
+        }
     }
 }
